[Refactor] Combine some files
